Add files via upload
This is a replication of W. Zhang and S. Luck experiment "Discrete fixed-resolution representations in visual working memory"is a PsychoPy experiment with an interactive color wheel and MATLAB code for analysing data and creating a mixture model of memory representation of several objects (Memtoolbox). It was made by Elena Lenchukova and Nadezhda Murzyakova, BA students of department of psycology, National Research University Higher School of Economics on an advanced course of cognitive psychology in 2018.